Create extra problem; clarify proof
In the proof that probability measures are monotone, the decomposition
$B = A \cup (B\setdiff A)$ for $A\subset B$ is used. However, to apply countable
additivity, it should be shown that both sets are in the $\sigma$-algebra.

Create an extra practice problem for the students to show that $\sigma$-algebras
are closed under countable intersection. That practice problem will be referred
to in the aforementioned proof.